requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Login==0.6.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
reportlab==4.2.5
